requirements.txt
cffi==1.17.1
cryptography==45.0.5
eralchemy==1.5.0
graphviz==0.21
greenlet==3.2.3
pycparser==2.22
PyMySQL==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
Flask==2.3.3
Flask-CORS==4.0.0
PyJWT==2.8.0
python-chess==1.999
openai==0.28.1
requests==2.31.0
python-dotenv==1.0.0
